requirements.txt
jupyter
pandas
awscrt==0.13.6
awsiotsdk==1.10.0